added popup.js
